plugins/computer_use/requirements.txt
opencv-python
numpy
Pillow
torch
torchvision
easyocr
rapidocr-onnxruntime
git+https://github.com/apple/ml-mobileclip.git
